+++ tests/client/karmaConf.js
@@ -7,6 +7,38 @@ module.exports = function(config) {
 
   config.set({
 
+    // Use mocha and chai for tests
+    frameworks: ['mocha', 'chai'],
+
+    // Web server port
+    port: 9876,
+
+    // Enable / disable colors in the output (reporters and logs)
+    colors: true,
+
+    // level of logging
+    // possible values: DISABLE || ERROR || WARN || INFO || DEBUG
+    logLevel: 'INFO',
+
+    // Enable / disable watching file and executing tests whenever
+    // any file changes
+    autoWatch: false,
+
+    // List of browsers to execute tests on
+    browsers: ['ChromeHeadlessCI'],
+
+    // Configure custom ChromHeadlessCI as an extension of ChromeHeadless without sandbox
+    customLaunchers: {
+      ChromeHeadlessCI: {
+        base: 'ChromeHeadless',
+        flags: ['--no-sandbox']
+      }
+    },
+
+    // Continuous Integration mode
+    // if true, Karma captures browsers, runs the tests and exits
+    singleRun: true,
+
     // Base path that will be used to resolve all patterns
     // (eg. files, exclude)
     basePath: '../../',